Fix test type detection
Currently if the `--debug` is not provided to `delfino test`, then an empty positional argument is passed to `pytest`.  This is interpreted as "current directory" by pytest, which means that each test type runs the full suite of tests available under the project root directory (assuming the command is run from the root).

This PR updates the pytest invocation to omit the `-s` argument if `--debug` is not supplied.